Player Sprint Animation
Created transition in the player's animator from player idle to sprint. Created the boolean parameter "sprint" inside of the animator.

In PlayerController.cs, created the Animator handle and 3 float values that will store the navMesh velocity.
If conditionals to check navMesh speed and if speed is not 0, then set the "sprint" parameter to true. If velocity is 0, set sprint parameter to false and that will handle the animator transition.